# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pulse Watch API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MonitoredEndpointStatus = typeof MonitoredEndpointStatus[keyof typeof MonitoredEndpointStatus];


export const MonitoredEndpointStatus = {
  healthy: 'healthy',
  warning: 'warning',
  failed: 'failed',
  pending: 'pending',
} as const;

export interface MonitoredEndpoint {
  id: number;
  name: string;
  curlCommand: string;
  expectedStatusCode: number;
  expectedResponseTime: number;
  status: MonitoredEndpointStatus;
  /** @nullable */
  lastResponseTime: number | null;
  /** @nullable */
  lastCheckedAt: string | null;
  createdAt: string;
}

export interface EndpointInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  curlCommand: string;
  expectedStatusCode: number;
  expectedResponseTime: number;
}

export type HistoryPointStatus = typeof HistoryPointStatus[keyof typeof HistoryPointStatus];


export const HistoryPointStatus = {
  success: 'success',
  error: 'error',
} as const;

export interface HistoryPoint {
  timestamp: string;
  responseTime: number;
  status: HistoryPointStatus;
}

export interface OverviewStats {
  totalApis: number;
  healthyApis: number;
  failedApis: number;
  warningApis: number;
  avgResponseTime: number;
}

export interface Execution {
  id: number;
  startedAt: string;
  totalApis: number;
  passed: number;
  failed: number;
  avgResponseTime: number;
}

export type ExecutionResultStatus = typeof ExecutionResultStatus[keyof typeof ExecutionResultStatus];


export const ExecutionResultStatus = {
  passed: 'passed',
  failed: 'failed',
} as const;

export interface ExecutionResult {
  id: number;
  executionId: number;
  endpointId: number;
  endpointName: string;
  endpointUrl: string;
  status: ExecutionResultStatus;
  responseTime: number;
  /** @nullable */
  errorMessage?: string | null;
}

export interface ExecutionDetail {
  id: number;
  startedAt: string;
  totalApis: number;
  passed: number;
  failed: number;
  avgResponseTime: number;
  results: ExecutionResult[];
}

export interface Settings {
  id: number;
  checkIntervalMinutes: number;
}

export interface SettingsInput {
  /**
     * @minimum 1
     * @maximum 60
     */
  checkIntervalMinutes: number;
}

export type TestResultStatus = typeof TestResultStatus[keyof typeof TestResultStatus];


export const TestResultStatus = {
  passed: 'passed',
  failed: 'failed',
} as const;

export interface TestResult {
  status: TestResultStatus;
  responseTime: number;
  /** @nullable */
  statusCode?: number | null;
  /** @nullable */
  errorMessage?: string | null;
}