# lib/api-zod/src/generated/types/endpointInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pulse Watch API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EndpointInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  curlCommand: string;
  expectedStatusCode: number;
  expectedResponseTime: number;
}