# lib/api-zod/src/generated/types/execution.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pulse Watch API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Execution {
  id: number;
  startedAt: string;
  totalApis: number;
  passed: number;
  failed: number;
  avgResponseTime: number;
}