# lib/api-zod/src/generated/types/executionResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pulse Watch API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExecutionResultStatus } from './executionResultStatus';

export interface ExecutionResult {
  id: number;
  executionId: number;
  endpointId: number;
  endpointName: string;
  endpointUrl: string;
  status: ExecutionResultStatus;
  responseTime: number;
  /** @nullable */
  errorMessage?: string | null;
}