# lib/api-zod/src/generated/types/historyPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pulse Watch API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HistoryPointStatus } from './historyPointStatus';

export interface HistoryPoint {
  timestamp: string;
  responseTime: number;
  status: HistoryPointStatus;
}