# lib/api-zod/src/generated/types/monitoredEndpoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pulse Watch API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MonitoredEndpointStatus } from './monitoredEndpointStatus';

export interface MonitoredEndpoint {
  id: number;
  name: string;
  curlCommand: string;
  expectedStatusCode: number;
  expectedResponseTime: number;
  status: MonitoredEndpointStatus;
  /** @nullable */
  lastResponseTime: number | null;
  /** @nullable */
  lastCheckedAt: string | null;
  createdAt: string;
}