# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pulse Watch API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  id: number;
  checkIntervalMinutes: number;
}